webpay-server/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
transbank-sdk==6.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
